monitor bugfix

+++ monitor/src_loader.py
@@ -4,6 +4,7 @@ import numpy as np
 import hashlib
 from champss_timing.database import database
 
+
 class src_loader():
     def __init__(self, source_dir):
         self.source_dir = source_dir
@@ -39,7 +40,7 @@ class src_loader():
         resids = timing_info["residuals"]
 
         return {"mjd": mjds, "val": resids["val"], "err": resids["err"]}
-    
+
     def get_parameter_info(self):
         all_timing_info = self.db.get_all_timing_info()
         mjd = []
@@ -53,12 +54,12 @@ class src_loader():
             mjd.append(max(timing_info["obs_mjds"]))
             for key in parameter_info.keys():
                 parameter_info[key].append(timing_info["fitted_params"][key])
-        
+
         for key in parameter_info.keys():
             parameter_info[key] = {"val": parameter_info[key], "mjd": mjd}
 
         return parameter_info
-    
+
     def get_statistics(self):
         all_timing_info = self.db.get_all_timing_info()
         mjd = []
@@ -68,20 +69,20 @@ class src_loader():
             mjd.append(max(timing_info["obs_mjds"]))
             statistics["chi2"].append(timing_info["chi2"])
             statistics["chi2r"].append(timing_info["chi2_reduced"])
-            statistics["rms"].append(np.sqrt(np.mean(np.square(timing_info["residuals"]["val"]))))
+            statistics["rms"].append(float(np.sqrt(np.mean(np.square(timing_info["residuals"]["val"])))))
 
         for key in statistics.keys():
             statistics[key] = {"val": statistics[key], "mjd": mjd}
 
         return statistics
-    
+
     def get_parfile(self):
         return self.db.get_last_timing_info()["notes"]["fitted_parfile"]
-    
+
     def get_last_updated(self):
-        # return in YYYY-MM-DD 
+        # return in YYYY-MM-DD
         return datetime.datetime.utcfromtimestamp(self.db.get_last_timing_info()["timestamp"]).strftime('%Y-%m-%d')
-    
+
     def get_diagnostic_pdf_base64(self):
         with open(self.pdf, "rb") as f:
             return base64.b64encode(f.read()).decode("utf-8")
